fix prompt
